This is challange three for echapter eight
Challenge 3
Write a script that creates an envelope polygon feature class for the
Hawaii.shp feature class. There is actually a tool that accomplishes this
called Minimum Bounding Geometry. You can look at the tool to get
some ideas, but your script needs to work directly with the geometry
properties