Challenge 1
Create a custom function called countstringfields that determines the
number of fields of type string in an input feature class. Create this
function in a separate script ( for example, mycount.py ) that you call
from another script ( for example, callingscript.py ). You can use the
streets.shp feature class in the Exercise12 folder.